Fix race condition on shutdown event
When the event loop receives the shutdown event the event loop task
tried to TerminateDestroy the mpv handle and subsequently wait for
itself to finish.
Instead of calling TerminateDestroy in the shutdown event handler, it
relies on the Mpv::Dispose(bool) function to do so.
And the event loops stop functions checks if it's called by it's own
thread to avoid waiting on itself.